ExtSpectSolo demo app, v1
The top level contains the ExtSpectSolo demo app for the ExtSpect
plugin. It lets you see how ExtSpect operates within the demo app.

The ExtSpect plugin is in the extspect/ directory. Look at
extspect/ReadMe.txt for further instructions.

For users of the demos app, ExtSpectSolo, the index.html file expects
that a touch2 directory will be a sibling of the app directory. Please
make adjustments in index.html as needed.